Made changes as per review comments
