+++ config/registry/schemas/opta-config-files/aws-service.json
@@ -55,7 +55,7 @@
             "$ref": "https://app.runx.dev/modules/helm-chart"
           },
           {
-            "$ref": "https://app.runx.dev/modules/k8s-resource"
+            "$ref": "https://app.runx.dev/modules/k8s-manifest"
           },
           {
             "$ref": "https://app.runx.dev/modules/lambda-function"
