+++ config/registry/schemas/opta-config-files/azure-service.json
@@ -28,7 +28,7 @@
             "$ref": "https://app.runx.dev/modules/helm-chart"
           },
           {
-            "$ref": "https://app.runx.dev/modules/k8s-resource"
+            "$ref": "https://app.runx.dev/modules/k8s-manifest"
           }
         ]
       }
