+++ config/registry/schemas/opta-config-files/gcp-service.json
@@ -37,7 +37,7 @@
             "$ref": "https://app.runx.dev/modules/helm-chart"
           },
           {
-            "$ref": "https://app.runx.dev/modules/k8s-resource"
+            "$ref": "https://app.runx.dev/modules/k8s-manifest"
           }
         ]
       }
