+++ config/registry/schemas/opta-config-files/local-service.json
@@ -19,7 +19,7 @@
             "$ref": "https://app.runx.dev/modules/helm-chart"
           },
           {
-            "$ref": "https://app.runx.dev/modules/k8s-resource"
+            "$ref": "https://app.runx.dev/modules/k8s-manifest"
           },
           {
             "$ref": "https://app.runx.dev/modules/local-k8s-service"
