+++ modules/k8s_manifest/k8s-manifest.json
@@ -1,5 +1,5 @@
 {
-  "$id": "https://app.runx.dev/modules/k8s-resource",
+  "$id": "https://app.runx.dev/modules/k8s-manifest",
   "description": "Kubernetes Resource",
   "type": "object",
   "properties": {
@@ -10,9 +10,9 @@
     "type": {
       "description": "The name of this module",
       "enum": [
-        "k8s-resource"
+        "k8s-manifest"
       ],
-      "default": "k8s-resource",
+      "default": "k8s-manifest",
       "type": "string"
     }
   },
@@ -24,8 +24,8 @@
       "gcp",
       "azure"
     ],
-    "name": "k8s-resource",
-    "display_name": "k8s-resource",
+    "name": "k8s-manifest",
+    "display_name": "k8s-manifest",
     "tags": []
   },
   "required": [
