+++ modules/k8s_manifest/tests/test_k8s_manifest.py
@@ -1,0 +1,20 @@
+import os
+
+
+from modules.k8s_manifest.k8s_manifest import K8smanifestProcessor
+from opta.layer import Layer
+
+
+class TestK8smanifestProcessor:
+    def test_local_getconfig(self) -> None:
+        layer = Layer.load_from_yaml(
+            os.path.join(
+                os.getcwd(), "tests", "fixtures", "dummy_data", "local_dummy_config.yaml"
+            ),
+            None,
+        )
+        idx = len(layer.modules)
+        k8smanifest_module = layer.get_module("k8smanifest", idx)
+        K8smanifestProcessor(k8smanifest_module, layer)
+        assert k8smanifest_module.data["kubeconfig"] == "~/.kube/config"
+        assert k8smanifest_module.data["kubecontext"] == "kind-opta-local-cluster"
