+++ tests/fixtures/dummy_data/gcp_dummy_config.yaml
@@ -4,8 +4,8 @@ environments:
     variables: {}
 name: gcp-dummy-config
 modules:
-  - name: k8sresource
-    type: k8s-resource
+  - name: k8smanifest
+    type: k8s-manifest
     file_path: k8test.yaml
   - name: database
     type: gcp-postgres
